Remove extra delete
Now that the custom resource handles clearing the s3 bucket, we no longer
need to do it manually and retry the stack deletion.  We had to make this
catch the errors from the commands and ignore them before because the bucket
could have been empty, but now we can remove that as well and just delete
the stack, if the command for that fails then the script fails too.

PR for custom resource: https://github.com/DataDog/serverless-remote-instrumentation/pull/162/