Automate cdk upgrade
Automate the cdk upgrade flow.  To do this I created a script to parsed
the cdk version out of the projen rc file, then check the npm version
from  1 month ago.  I chose that because that's the minimum version that
someone needs, so 1 month behind seemed like a middle ground between being
recent enough and not being so aggressive that someone is constantly having
to upgrade cdk because of us.  Let me know if you prefer a different value
for this.  If they don't match, update it.

I tested this locally and it seems to work, I'll run it manually once
this is merged to know for sure.

The script has to be included in the github upgrade workflow, so we have
to overwrite it in projen since its managed for us.

Issues: https://github.com/DataDog/datadog-cdk-constructs/issues/465

Projen docs about this: https://projen.io/docs/project-types/aws-cdk-construct-library/#depending-on-aws-cdk-lib