Add tests and lint to cdk constructs release process
Add the test and lint step to the CDK constructs release pipeline.  Also include
a way to manually trigger this while we're still working on getting this setup.

One thing I haven't been able to figure out is how to test this without PRing and
merging it in, when we make a pipeline in the cdk constructs repo it uses the
version of this that is on main, and the config for that is in a different repo
that takes days to make changes to.  If anyone has an idea for how to test this
pre merge, I'd love to hear it.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-7287