Add github login script
Add a github login script and add a test run of it into the pipeline. The script
expects a vault path, and for it to have an app ID, installation ID, and key in
the vault secrets.  This script can be used for other repos that need to be logged
in to github to do releases.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-7287