Allow unit tests to run in gitlab (#471)
Allow the unit tests to work in gitlab.  There is something about our tests
that is probably getting re used when we may not intend to, or something that
isn't getting reset properly.  This was making the tests hang in gitlab, I'm
not sure why but for now lets set run in band to make it run sequentially.
This takes longer but is consistent.

Also change the github URL to match either the github or the gitlab url, with
the .git at the end being optional.

Modify the names of the example stacks os that they don't overlap as well,
this was giving us a warning.

* Passing gitlab job: https://gitlab.ddbuild.io/DataDog/datadog-cdk-constructs/-/jobs/1079139184
* Github / Gitlab issue: https://gitlab.ddbuild.io/DataDog/datadog-cdk-constructs/-/jobs/1079119360#L317
* Hanging gitlab job: https://gitlab.ddbuild.io/DataDog/datadog-cdk-constructs/-/jobs/1079015266

JIRA: https://datadoghq.atlassian.net/browse/SVLS-7287