Add pypi and npm login scripts, with cdk constructs release script
Add a script to log into npm and into pypi.  I've tested them locally and they work without
logging any credentials.  Create a release script for cdk constructs that just does the login
for those 3 and runs after the test / lint steps.

Also some minor changes to the github script to not log the extra running locally line, to
fix the usage text when called from another script.

This is a job where it ran, though it was in an infinite loop because of how I was testing it.
* https://gitlab.ddbuild.io/DataDog/datadog-cdk-constructs/-/jobs/1082738208

JIRA: https://datadoghq.atlassian.net/browse/SVLS-7287