chore: Add create-release script (#473)
* Add create-release script

Add a create release script to run when we want to make a release.  I left
this in this repo (at least for now) since this only does a commit / tag
and push, and didn't seem worth making us symlink the scripts folder in
the ci repo.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-7287

* Update scripts/create_release.sh

Co-authored-by: Tal Usvyatsky <18072279+TalUsvyatsky@users.noreply.github.com>

* Apply suggestions from code review

* Update scripts/create_release.sh

* Update scripts/create_release.sh

---------

Co-authored-by: Tal Usvyatsky <18072279+TalUsvyatsky@users.noreply.github.com>