Add version.json update to create release script
Add a line to update the version.json file in the create_release script.
This is needed both to keep it up to date, and to allow us to use a non
empty PR to approve releases.

Also add chore: to the prefix of the commit so it gets added to the
PR title to keep the PR title linter happy.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-7287