Move yarn standard-version
Remove the yarn standard-version command, since what we're doing with
that is mostly already done in the create release command.  I'm going
to probably copy this command as is to run in that instead of doing the
tags ourselves, but that'll be in a different PR since it's in a different
repo.

JIRA: https://datadoghq.atlassian.net/browse/SVLS-7287